\name{measuRing-package}
\alias{measuRing-package}
\alias{measuRing}
\docType{package}
\title{Detection and Control of Tree-Ring Widths on Scanned Image Sections}
\description{Identification of ring borders on scanned image sections from dendrochronological samples. Processesing of image reflectances to produce gray matrices and a time series of smoothed gray values. Luminance data is plotted on segmented images for users to perform both: visual identification of ring borders, or control of automatic detection. Routines to visually include/exclude ring borders on the R graphical device, or automatically detect ring borders using a linear detection algorithm. This algorithm detects ring borders according to positive/negative extreme values in the smoothed time-series of gray values.}
\details{
\tabular{ll}{Package: \tab measuRing\cr
Type: \tab Package\cr
Title: \tab Detection and Control of Tree-Ring Widths on Scanned Image Sections\cr
Version: \tab 0.3\cr
Author: \tab Wilson Lara, Carlos Sierra, Felipe Bravo\cr
Date: \tab 2015-03-23\cr
Maintainer: \tab Wilson Lara <wilarhen@gmail.com>\cr
Depends: \tab pastecs, png, tiff\cr
License: \tab GPL-3\cr
LazyData: \tab TRUE\cr}
}
\author{Wilson Lara, Carlos Sierra, Felipe Bravo}

\keyword{ package }


