\name{measuRing-package}
\alias{measuRing-package}
\alias{measuRing}
\docType{package}
\title{Detection and Control of Tree-Ring Widths on Scanned Image Sections}

\description{Identification of ring borders on scanned image sections from dendrochronological samples. Processing of image reflectances to produce gray matrices and time series of smoothed gray values. Luminance data is plotted on segmented images for users to perform both: visual identification of ring borders, or control of automatic detection. Routines to visually include/exclude ring borders on the R graphical device, or automatically detect ring borders using a linear detection algorithm. This algorithm detects ring borders according to positive/negative extreme values in the smoothed time-series of gray values.}

\details{

The DESCRIPTION file:
\packageDESCRIPTION{measuRing}
\packageIndices{measuRing}
Maintainer: \packageMaintainer{measuRing}
}
\author{Wilson Lara, Carlos Sierra, Felipe Bravo}



\keyword{ package }


